+++ lib/td_dd/data_structures/data_structure_type.ex
@@ -6,18 +6,17 @@ defmodule TdDd.DataStructures.DataStructureType do
   import Ecto.Changeset
 
   schema "data_structure_types" do
-    field :structure_type, :string
-    field :template_id, :integer
-    field :translation, :string
-    field :metadata_fields, :map
+    field(:structure_type, :string)
+    field(:template_id, :integer)
+    field(:translation, :string)
+    field(:metadata_fields, :map)
 
     timestamps()
   end
 
-  @doc false
-  def changeset(data_structure_type, attrs) do
+  def changeset(data_structure_type, params) do
     data_structure_type
-    |> cast(attrs, [:structure_type, :translation, :template_id, :metadata_fields])
+    |> cast(params, [:structure_type, :translation, :template_id, :metadata_fields])
     |> validate_required([:structure_type, :template_id])
     |> unique_constraint(:structure_type)
   end