+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ## [Unreleased]
 
-### Changed
+### Fixed
 
-- [TD-2966] Increased timeout in `StructureLoader.refresh/1`
+- [TD-2979] Timeout issues loading metadata
 
 ## [4.3.0] 2020-09-07